day05/requirements.txt
asgiref>=3.3.4
autopep8>=1.5.7
Django>=4.0.4
psycopg2-binary>=2.8.0
pycodestyle>=2.7.0
pytz>=2021.1
sqlparse>=0.4.1
toml>=0.10.2